refactor: move withErrorBoundary HoC into util folder
